+++ micromouse/IRSensor.h
@@ -36,7 +36,7 @@ namespace Micromouse
 	private:
 		void initPins();
 		void defaultCalibration();
-		int interpolate(int x1, int x2, int y1, int y2, int x);
+		float interpolate(int x1, int x2, int y1, int y2, int x);
 
 		int test = 0;
